extract map to its own component
